+++ src/main/java/com/codenear/butterfly/address/domain/repository/AddressRepository.java
@@ -1,5 +1,6 @@
-package com.codenear.butterfly.address.domain;
+package com.codenear.butterfly.address.domain.repository;
 
+import com.codenear.butterfly.address.domain.Address;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.LinkedList;
