+++ src/main/java/com/codenear/butterfly/payment/kakaoPay/application/CancelPaymentService.java
@@ -1,6 +1,6 @@
 package com.codenear.butterfly.payment.kakaoPay.application;
 
-import com.codenear.butterfly.address.domain.AddressRepository;
+import com.codenear.butterfly.address.domain.repository.AddressRepository;
 import com.codenear.butterfly.member.domain.repository.member.MemberRepository;
 import com.codenear.butterfly.notify.fcm.application.FCMFacade;
 import com.codenear.butterfly.payment.application.PaymentService;
